impr: allow bypassing slow launch work
